Livin' on the Edge keys upgrade (#38)
I tried to switch over to the high strings, but they were too hard to chart, especially when they got out of alignment with the tempo map after the piano stopped. Since I can't change the tempo map, I just stuck with the piano until 109 and switched over to low strings there.

Overdrive was also a pain because of the imbalance of phrases between instruments: 15 for drums, 7 for bass, 6 for guitar, and 9 for vox. I put in 7 phrases for keys, but let me know if I should add one more.